error with the fetch data function
