+++ frontend/src/app.js
@@ -4,21 +4,22 @@ import Home from './pages/Home';
 import Login from './pages/Login';
 import Signup from './pages/Signup';
 import Profile from './pages/Profile';
-import UserData from './components/UserData';
 import CurrentUserProvider from './context/CurrentUser';
+
+
 
 
 function App() {
   return (
     <CurrentUserProvider>
-      <Routes>
+     <Routes>
         <Route exact path='/' element={<Home/>}/>
         <Route exact path='/login' element={<Login/>}/>
         <Route exact path='/signup' element={<Signup/>}/>
         <Route exact path='/profile' element={<Profile/>}/>
-        <Route exact path='/userdata' element={<UserData/>}/>
       </Routes>
     </CurrentUserProvider>
+    
   );
 }
